requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0 